renamed dockless directory and references
