+++ _layouts/homepage-layout.html
@@ -116,7 +116,7 @@
 
         <div class="col-lg-7">  <!-- buttons column -->
           <div class="row">
-            <a class="nav-homepage col-md-4 col-xl-3 col-sm-6 p-1" href='{{ site.baseurl }}/dockless-data'>
+            <a class="nav-homepage col-md-4 col-xl-3 col-sm-6 p-1" href='{{ site.baseurl }}/micromobility-data'>
               <div class="col homepage-button homepage-button-row-3 p-2">
                 <h5>Shared Micromobility Data</h5>
                 <i  class="fa button-icon fa-bolt"></i>
