+++ _layouts/micromobility-card-layout.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+  <html lang="en-us">
+
+    {% include head.html %}
+
+    <body>
+
+        {% include coa-header.html %}
+
+        {% include dash-top-nav.html %}
+
+      <div class="container-fluid">
+
+        <div class="row dash-panel-row no-gutters" id="panel-row-micromobility">
+        <!-- cards go here programmatically -->
+        </div>
+
+      </div> <!-- end container -->
+
+      <div id="root-MicromobilityData"></div>
+
+
+    {% include footer.html %}
+
+    {% include default-scripts.html %}
+
+    {% if page.custom_js %}
+
+      {% for js_file in page.custom_js %}
+
+        <script src='{{ site.baseurl }}/webpackjs/output/{{ js_file }}.js' type="text/javascript"></script>
+
+      {% endfor %}
+
+    {% endif %}
+
+
+  </body>
+</html>
