--- dockless-data.md
@@ -1,10 +1,0 @@
----
-layout: dockless-card-layout
-title: Shared Micromobility Overview | Transportation Data and Performance Hub
-title_short: Shared Micromobility Overview
-description: Shared Micromobility Overview
-custom_js:
-- dockless.bundle
-custom_css:
-- dashboard
----
