+++ micromobility-data.md
@@ -1,0 +1,10 @@
+---
+layout: micromobility-card-layout
+title: Shared Micromobility Overview | Transportation Data and Performance Hub
+title_short: Shared Micromobility Overview
+description: Shared Micromobility Overview
+custom_js:
+- micromobility.bundle
+custom_css:
+- dashboard
+---
