--- webpackjs/components/DocklessData/CardContainer.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const CardContainer = ({ data, children }) => (
-  <div className="row dash-panel-row no-gutters">{children}</div>
-);
-
-export default CardContainer;
