+++ webpackjs/components/MicromobilityData/CardContainer.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const CardContainer = ({ data, children }) => (
+  <div className="row dash-panel-row no-gutters">{children}</div>
+);
+
+export default CardContainer;
